Core module shared classes refactoring
Changing mtehid parameters name, moving specific packages from core
