--- modules/core/shared/src/main/scala/com/github/fsanaulla/chronicler/core/encoding/package.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright 2017-2019 Faiaz Sanaulla
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.fsanaulla.chronicler.core
-
-package object encoding {
-
-  // extract encoding from content type
-  def encodingFromContentType(ct: String): Option[String] =
-    ct.split(";")
-      .map(_.trim.toLowerCase)
-      .collectFirst {
-        case s if s.startsWith("charset=") && s.substring(8).trim != "" => s.substring(8).trim
-      }
-}
